// apps/web/src/lib/enums.ts
// Entities that have a create/edit/import/delete drawer.
export enum DrawerEntity {
  FIELD_TRANSFORMATION = "field-transformation",
  TRANSFORMATION_RULE = "transformation-rule",
  DISPATCH_TYPE = "dispatch-type",
  VIEW_TOKEN = "view-token",
}

export enum DrawerMode {
  CREATE = "create",
  EDIT = "edit",
  IMPORT = "import",
  DELETE = "delete",
}